refactor(entry): extract main() into orchestrator, preflight, export_flow
- main.rs: 913 → 88 lines (thin entry point: parse CLI → delegate)
- New orchestrator.rs: 647 lines (full pipeline orchestration)
- New preflight.rs: 189 lines (config merge, HTTP connectivity, helpers)
- New export_flow.rs: 192 lines (RAG export + AI cleaning + file save)
- All 448 tests passing, both default and --features ai compile
